Move map, template, tile/set & property parse fns
