+++ src/map.rs
@@ -7,10 +7,11 @@ use xml::attribute::OwnedAttribute;
 use crate::{
     error::{Error, Result},
     layers::{LayerData, LayerTag},
-    properties::{parse_properties, Color, Properties},
+    parse::common::tileset::EmbeddedParseResultType,
+    properties::{Color, Properties},
     tileset::Tileset,
     util::{get_attrs, parse_tag, XmlEventResult},
-    EmbeddedParseResultType, Layer, ResourceCache, ResourceReader,
+    Layer, ResourceCache, ResourceReader,
 };
 
 pub(crate) struct MapTilesetGid {
@@ -21,7 +22,7 @@ pub(crate) struct MapTilesetGid {
 /// All Tiled map files will be parsed into this. Holds all the layers and tilesets.
 #[derive(PartialEq, Clone, Debug)]
 pub struct Map {
-    version: String,
+    pub(crate) version: String,
     /// The way tiles are laid out in the map.
     pub orientation: Orientation,
     /// Width of the map, in tiles.
@@ -49,14 +50,14 @@ pub struct Map {
     /// will be the same as the one from the tilesets the map is using.
     pub tile_height: u32,
     /// The tilesets present on this map.
-    tilesets: Vec<Arc<Tileset>>,
+    pub(crate) tilesets: Vec<Arc<Tileset>>,
     /// The layers present in this map.
-    layers: Vec<LayerData>,
+    pub(crate) layers: Vec<LayerData>,
     /// The custom properties of this map.
     pub properties: Properties,
     /// The background color of this map, if any.
     pub background_color: Option<Color>,
-    infinite: bool,
+    pub(crate) infinite: bool,
 }
 
 impl Map {
@@ -115,139 +116,6 @@ impl Map {
     /// Returns the layer that has the specified index, if it exists.
     pub fn get_layer(&self, index: usize) -> Option<Layer> {
         self.layers.get(index).map(|data| Layer::new(self, data))
-    }
-}
-
-impl Map {
-    pub(crate) fn parse_xml(
-        parser: &mut impl Iterator<Item = XmlEventResult>,
-        attrs: Vec<OwnedAttribute>,
-        map_path: &Path,
-        reader: &mut impl ResourceReader,
-        cache: &mut impl ResourceCache,
-    ) -> Result<Map> {
-        let ((c, infinite), (v, o, w, h, tw, th)) = get_attrs!(
-            for v in attrs {
-                Some("backgroundcolor") => colour ?= v.parse(),
-                Some("infinite") => infinite = v == "1",
-                "version" => version = v,
-                "orientation" => orientation ?= v.parse::<Orientation>(),
-                "width" => width ?= v.parse::<u32>(),
-                "height" => height ?= v.parse::<u32>(),
-                "tilewidth" => tile_width ?= v.parse::<u32>(),
-                "tileheight" => tile_height ?= v.parse::<u32>(),
-            }
-            ((colour, infinite), (version, orientation, width, height, tile_width, tile_height))
-        );
-
-        let infinite = infinite.unwrap_or(false);
-
-        // We can only parse sequentally, but tilesets are guaranteed to appear before layers.
-        // So we can pass in tileset data to layer construction without worrying about unfinished
-        // data usage.
-        let mut layers = Vec::new();
-        let mut properties = HashMap::new();
-        let mut tilesets = Vec::new();
-
-        parse_tag!(parser, "map", {
-            "tileset" => |attrs: Vec<OwnedAttribute>| {
-                let res = Tileset::parse_xml_in_map(parser, &attrs, map_path,  reader, cache)?;
-                match res.result_type {
-                    EmbeddedParseResultType::ExternalReference { tileset_path } => {
-                        let tileset = if let Some(ts) = cache.get_tileset(&tileset_path) {
-                            ts
-                        } else {
-                            let tileset = Arc::new(crate::parse::xml::parse_tileset(&tileset_path,  reader, cache)?);
-                            cache.insert_tileset(tileset_path.clone(), tileset.clone());
-                            tileset
-                        };
-
-                        tilesets.push(MapTilesetGid{first_gid: res.first_gid, tileset});
-                    }
-                    EmbeddedParseResultType::Embedded { tileset } => {
-                        tilesets.push(MapTilesetGid{first_gid: res.first_gid, tileset: Arc::new(tileset)});
-                    },
-                };
-                Ok(())
-            },
-            "layer" => |attrs| {
-                layers.push(LayerData::new(
-                    parser,
-                    attrs,
-                    LayerTag::Tiles,
-                    infinite,
-                    map_path,
-                    &tilesets,
-                    None,
-                    reader,
-                    cache
-                )?);
-                Ok(())
-            },
-            "imagelayer" => |attrs| {
-                layers.push(LayerData::new(
-                    parser,
-                    attrs,
-                    LayerTag::Image,
-                    infinite,
-                    map_path,
-                    &tilesets,
-                    None,
-                    reader,
-                    cache
-                )?);
-                Ok(())
-            },
-            "objectgroup" => |attrs| {
-                layers.push(LayerData::new(
-                    parser,
-                    attrs,
-                    LayerTag::Objects,
-                    infinite,
-                    map_path,
-                    &tilesets,
-                    None,
-                    reader,
-                    cache
-                )?);
-                Ok(())
-            },
-            "group" => |attrs| {
-                layers.push(LayerData::new(
-                    parser,
-                    attrs,
-                    LayerTag::Group,
-                    infinite,
-                    map_path,
-                    &tilesets,
-                    None,
-                    reader,
-                    cache
-                )?);
-                Ok(())
-            },
-            "properties" => |_| {
-                properties = parse_properties(parser)?;
-                Ok(())
-            },
-        });
-
-        // We do not need first GIDs any more
-        let tilesets = tilesets.into_iter().map(|ts| ts.tileset).collect();
-
-        Ok(Map {
-            version: v,
-            orientation: o,
-            width: w,
-            height: h,
-            tile_width: tw,
-            tile_height: th,
-            tilesets,
-            layers,
-            properties,
-            background_color: c,
-            infinite,
-        })
     }
 }
