+++ src/parse/mod.rs
@@ -1,1 +1,2 @@
 pub mod xml;
+pub mod common;
