+++ src/parse/xml/mod.rs
@@ -1,4 +1,5 @@
-mod map;
-pub use map::*;
-mod tileset;
-pub use tileset::*;
+pub mod map;
+pub mod properties;
+pub mod template;
+pub mod tile;
+pub mod tileset;
