+++ src/properties.rs
@@ -84,83 +84,5 @@ pub enum PropertyValue {
     ObjectValue(u32),
 }
 
-impl PropertyValue {
-    fn new(property_type: String, value: String) -> Result<PropertyValue> {
-        // Check the property type against the value.
-        match property_type.as_str() {
-            "bool" => match value.parse() {
-                Ok(val) => Ok(PropertyValue::BoolValue(val)),
-                Err(err) => Err(Error::InvalidPropertyValue {
-                    description: err.to_string(),
-                }),
-            },
-            "float" => match value.parse() {
-                Ok(val) => Ok(PropertyValue::FloatValue(val)),
-                Err(err) => Err(Error::InvalidPropertyValue {
-                    description: err.to_string(),
-                }),
-            },
-            "int" => match value.parse() {
-                Ok(val) => Ok(PropertyValue::IntValue(val)),
-                Err(err) => Err(Error::InvalidPropertyValue {
-                    description: err.to_string(),
-                }),
-            },
-            "color" if value.len() > 1 => Color::from_str(&value)
-                .map(PropertyValue::ColorValue)
-                .map_err(|_| Error::InvalidPropertyValue {
-                    description: "Couldn't parse color".to_string(),
-                }),
-            "string" => Ok(PropertyValue::StringValue(value)),
-            "object" => match value.parse() {
-                Ok(val) => Ok(PropertyValue::ObjectValue(val)),
-                Err(err) => Err(Error::InvalidPropertyValue {
-                    description: err.to_string(),
-                }),
-            },
-            "file" => Ok(PropertyValue::FileValue(value)),
-            _ => Err(Error::UnknownPropertyType {
-                type_name: property_type,
-            }),
-        }
-    }
-}
-
 /// A custom property container.
 pub type Properties = HashMap<String, PropertyValue>;
-
-pub(crate) fn parse_properties(
-    parser: &mut impl Iterator<Item = XmlEventResult>,
-) -> Result<Properties> {
-    let mut p = HashMap::new();
-    parse_tag!(parser, "properties", {
-        "property" => |attrs:Vec<OwnedAttribute>| {
-            let (t, v_attr, k) = get_attrs!(
-                for attr in attrs {
-                    Some("type") => obj_type = attr,
-                    Some("value") => value = attr,
-                    "name" => name = attr
-                }
-                (obj_type, value, name)
-            );
-            let t = t.unwrap_or_else(|| "string".to_owned());
-
-            let v: String = match v_attr {
-                Some(val) => val,
-                None => {
-                    // if the "value" attribute was missing, might be a multiline string
-                    match parser.next() {
-                        Some(Ok(XmlEvent::Characters(s))) => Ok(s),
-                        Some(Err(err)) => Err(Error::XmlDecodingError(err)),
-                        None => unreachable!(), // EndDocument or error must come first
-                        _ => Err(Error::MalformedAttributes(format!("property '{}' is missing a value", k))),
-                    }?
-                }
-            };
-
-            p.insert(k, PropertyValue::new(t, v)?);
-            Ok(())
-        },
-    });
-    Ok(p)
-}
