+++ src/tileset.rs
@@ -5,7 +5,7 @@ use xml::attribute::OwnedAttribute;
 
 use crate::error::{Error, Result};
 use crate::image::Image;
-use crate::properties::{parse_properties, Properties};
+use crate::properties::Properties;
 use crate::tile::TileData;
 use crate::{util::*, Gid, ResourceCache, ResourceReader, Tile, TileId};
 
@@ -53,36 +53,13 @@ pub struct Tileset {
     pub image: Option<Image>,
 
     /// All the tiles present in this tileset, indexed by their local IDs.
-    tiles: HashMap<TileId, TileData>,
+    pub(crate) tiles: HashMap<TileId, TileData>,
 
     /// All the wangsets present in this tileset.
     pub wang_sets: Vec<WangSet>,
 
     /// The custom properties of the tileset.
     pub properties: Properties,
-}
-
-pub(crate) enum EmbeddedParseResultType {
-    ExternalReference { tileset_path: PathBuf },
-    Embedded { tileset: Tileset },
-}
-
-pub(crate) struct EmbeddedParseResult {
-    pub first_gid: Gid,
-    pub result_type: EmbeddedParseResultType,
-}
-
-/// Internal structure for holding mid-parse information.
-struct TilesetProperties {
-    spacing: Option<u32>,
-    margin: Option<u32>,
-    tilecount: u32,
-    columns: Option<u32>,
-    name: String,
-    tile_width: u32,
-    tile_height: u32,
-    /// The root all non-absolute paths contained within the tileset are relative to.
-    root_path: PathBuf,
 }
 
 impl Tileset {
@@ -98,206 +75,5 @@ impl Tileset {
         self.tiles
             .iter()
             .map(move |(id, data)| (*id, Tile::new(self, data)))
-    }
-}
-
-impl Tileset {
-    pub(crate) fn parse_xml_in_map(
-        parser: &mut impl Iterator<Item = XmlEventResult>,
-        attrs: &[OwnedAttribute],
-        path: &Path, // Template or Map file
-        reader: &mut impl ResourceReader,
-        cache: &mut impl ResourceCache,
-    ) -> Result<EmbeddedParseResult> {
-        Tileset::parse_xml_embedded(parser, attrs, path, reader, cache).or_else(|err| {
-            if matches!(err, Error::MalformedAttributes(_)) {
-                Tileset::parse_xml_reference(attrs, path)
-            } else {
-                Err(err)
-            }
-        })
-    }
-
-    fn parse_xml_embedded(
-        parser: &mut impl Iterator<Item = XmlEventResult>,
-        attrs: &[OwnedAttribute],
-        path: &Path, // Template or Map file
-        reader: &mut impl ResourceReader,
-        cache: &mut impl ResourceCache,
-    ) -> Result<EmbeddedParseResult> {
-        let ((spacing, margin, columns, name), (tilecount, first_gid, tile_width, tile_height)) = get_attrs!(
-           for v in attrs {
-            Some("spacing") => spacing ?= v.parse(),
-            Some("margin") => margin ?= v.parse(),
-            Some("columns") => columns ?= v.parse(),
-            Some("name") => name = v,
-            "tilecount" => tilecount ?= v.parse::<u32>(),
-            "firstgid" => first_gid ?= v.parse::<u32>().map(Gid),
-            "tilewidth" => tile_width ?= v.parse::<u32>(),
-            "tileheight" => tile_height ?= v.parse::<u32>(),
-           }
-           ((spacing, margin, columns, name), (tilecount, first_gid, tile_width, tile_height))
-        );
-
-        let root_path = path.parent().ok_or(Error::PathIsNotFile)?.to_owned();
-
-        Self::finish_parsing_xml(
-            parser,
-            TilesetProperties {
-                spacing,
-                margin,
-                name: name.unwrap_or_default(),
-                root_path,
-                columns,
-                tilecount,
-                tile_height,
-                tile_width,
-            },
-            reader,
-            cache,
-        )
-        .map(|tileset| EmbeddedParseResult {
-            first_gid,
-            result_type: EmbeddedParseResultType::Embedded { tileset },
-        })
-    }
-
-    fn parse_xml_reference(
-        attrs: &[OwnedAttribute],
-        map_path: &Path,
-    ) -> Result<EmbeddedParseResult> {
-        let (first_gid, source) = get_attrs!(
-            for v in attrs {
-                "firstgid" => first_gid ?= v.parse::<u32>().map(Gid),
-                "source" => source = v,
-            }
-            (first_gid, source)
-        );
-
-        let tileset_path = map_path.parent().ok_or(Error::PathIsNotFile)?.join(source);
-
-        Ok(EmbeddedParseResult {
-            first_gid,
-            result_type: EmbeddedParseResultType::ExternalReference { tileset_path },
-        })
-    }
-
-    pub(crate) fn parse_external_tileset(
-        parser: &mut impl Iterator<Item = XmlEventResult>,
-        attrs: &[OwnedAttribute],
-        path: &Path,
-        reader: &mut impl ResourceReader,
-        cache: &mut impl ResourceCache,
-    ) -> Result<Tileset> {
-        let ((spacing, margin, columns, name), (tilecount, tile_width, tile_height)) = get_attrs!(
-            for v in attrs {
-                Some("spacing") => spacing ?= v.parse(),
-                Some("margin") => margin ?= v.parse(),
-                Some("columns") => columns ?= v.parse(),
-                Some("name") => name = v,
-
-                "tilecount" => tilecount ?= v.parse::<u32>(),
-                "tilewidth" => tile_width ?= v.parse::<u32>(),
-                "tileheight" => tile_height ?= v.parse::<u32>(),
-            }
-            ((spacing, margin, columns, name), (tilecount, tile_width, tile_height))
-        );
-
-        let root_path = path.parent().ok_or(Error::PathIsNotFile)?.to_owned();
-
-        Self::finish_parsing_xml(
-            parser,
-            TilesetProperties {
-                spacing,
-                margin,
-                name: name.unwrap_or_default(),
-                root_path,
-                columns,
-                tilecount,
-                tile_height,
-                tile_width,
-            },
-            reader,
-            cache,
-        )
-    }
-
-    fn finish_parsing_xml(
-        parser: &mut impl Iterator<Item = XmlEventResult>,
-        prop: TilesetProperties,
-        reader: &mut impl ResourceReader,
-        cache: &mut impl ResourceCache,
-    ) -> Result<Tileset> {
-        let mut image = Option::None;
-        let mut tiles = HashMap::with_capacity(prop.tilecount as usize);
-        let mut properties = HashMap::new();
-        let mut wang_sets = Vec::new();
-
-        parse_tag!(parser, "tileset", {
-            "image" => |attrs| {
-                image = Some(Image::new(parser, attrs, &prop.root_path)?);
-                Ok(())
-            },
-            "properties" => |_| {
-                properties = parse_properties(parser)?;
-                Ok(())
-            },
-            "tile" => |attrs| {
-                let (id, tile) = TileData::new(parser, attrs, &prop.root_path, reader, cache)?;
-                tiles.insert(id, tile);
-                Ok(())
-            },
-            "wangset" => |attrs| {
-                let set = WangSet::new(parser, attrs)?;
-                wang_sets.push(set);
-                Ok(())
-            },
-        });
-
-        // A tileset is considered an image collection tileset if there is no image attribute (because its tiles do).
-        let is_image_collection_tileset = image.is_none();
-
-        if !is_image_collection_tileset {
-            for tile_id in 0..prop.tilecount {
-                tiles.entry(tile_id).or_default();
-            }
-        }
-
-        let margin = prop.margin.unwrap_or(0);
-        let spacing = prop.spacing.unwrap_or(0);
-        let columns = prop
-            .columns
-            .map(Ok)
-            .unwrap_or_else(|| Self::calculate_columns(&image, prop.tile_width, margin, spacing))?;
-
-        Ok(Tileset {
-            name: prop.name,
-            tile_width: prop.tile_width,
-            tile_height: prop.tile_height,
-            spacing,
-            margin,
-            columns,
-            tilecount: prop.tilecount,
-            image,
-            tiles,
-            wang_sets,
-            properties,
-        })
-    }
-
-    fn calculate_columns(
-        image: &Option<Image>,
-        tile_width: u32,
-        margin: u32,
-        spacing: u32,
-    ) -> Result<u32> {
-        image
-            .as_ref()
-            .map(|image| (image.width as u32 - margin + spacing) / (tile_width + spacing))
-            .ok_or_else(|| {
-                Error::MalformedAttributes(
-                    "No <image> nor columns attribute in <tileset>".to_string(),
-                )
-            })
     }
 }
